Get js search working
Still a lot to do, but it at least 'works'
