+++ wdn/templates_5.0/includes/global/search.html
@@ -55,8 +55,8 @@
         </form>
         </div>
 
-      <div class="dcf-search-results-wrapper dcf-h-100% dcf-overflow-y-scroll">
-        <div class="dcf-search-results dcf-grid-halves@md dcf-col-gap-vw dcf-row-gap-lg dcf-pt-6 dcf-pr-6 dcf-pb-10 dcf-txt-sm unl-font-sans">
+      <div id="dcf-search-results-wrapper" class="dcf-search-results-wrapper dcf-h-100% dcf-overflow-y-scroll">
+        <div style="display:none" class="dcf-search-results dcf-grid-halves@md dcf-col-gap-vw dcf-row-gap-lg dcf-pt-6 dcf-pr-6 dcf-pb-10 dcf-txt-sm unl-font-sans">
 
           <!-- Web -->
           <div class="dcf-search-results-web">
