+++ wdn/templates_5.0/js-src/main-wdn-plugins.babel.js
@@ -2,7 +2,7 @@
 
 define(['analytics', //this also loads the IDM plugin
 //'navigation', //TODO: do we need navigation js?
-//'search', //TODO: add search
+'search', //TODO: add search
 'unlalert'
 //'images', //TODO: do we need 'images'?
 //'form_validation', //TODO: do we need to init form_validation here?
